refactor(remediate,correctness): split the verification call and the lockfile check out of the runners
Both runners crossed the large-file floor with the 4.4.5 additions. The
lockfile-sync executor moves to src/analyzers/correctness/lockfile-check.ts
(still inside the canonical runner directory the arch-check allows), and
the remediate lane's call into verifyTree plus the install-failed note move
to src/remediate/verify.ts. The other remediate test suites gain the
worktree + install seams the runner now needs.
